victron/data/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
paho-mqtt==1.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
six==1.16.0
urllib3==2.0.2
